Putting exercisePipelinesVariables into good use. The Pipelines configuration now may be saved.
