+++ src/locales/cs.json
@@ -182,6 +182,19 @@
   "app.editAssignmentForm.validation.pointsPercentualThresholdMustBeInteger": "Procentuální limit bodů musí být celé číslo.",
   "app.editAssignmentForm.validation.sameLocalizedTexts": "Je vyplněno více jazykových variant pro jednu lokalizaci. Prosím ujistěte se, že lokalizace jsou unikátní.",
   "app.editAssignmentForm.validation.submissionsCountLimit": "Prosíme vyplňte maximální počet odevzdaných řešení jako kladné celé číslo.",
+  "app.editEnvironmentConfig.noRuntimeSelected": "Aby bylo možné pokračovat se zbytkem konfigurace, nejprve je třeba nastavit běhové prostředí.",
+  "app.editEnvironmentConfig.noVariables": "Zatím nejsou nastavené žádné proměnné ...",
+  "app.editEnvironmentConfig.selectedRuntime": "Vybrané běhové prostředí:",
+  "app.editEnvironmentConfig.selectedRuntimeInfo": "V pokročilé konfiguraci určuje vybrané běhové prostředí pouze požadavky kladené na vyhodnocovací server (požadované kompilátory, nástroje nebo knihovny). Ostatní věci (předpis odevzdávaných souborů, pipelines) je možné konfigurovat nezávisle.",
+  "app.editEnvironmentConfig.setDefaultVariables": "Nastavit výchozí proměnné",
+  "app.editEnvironmentConfig.title": "Konfigurace běhového prostředí",
+  "app.editEnvironmentConfig.tooltip.add": "Přidat další proměnnou.",
+  "app.editEnvironmentConfig.tooltip.remove": "Odstranit tuto proměnnou.",
+  "app.editEnvironmentConfig.validateName": "Toto není platný název proměnné.",
+  "app.editEnvironmentConfig.validateWildcard": "Tato hodnota neodpovídá názvu souboru ani wildcard formátu.",
+  "app.editEnvironmentConfig.variableName": "Proměnná vstupních souborů:",
+  "app.editEnvironmentConfig.variableValue": "Předpis názvu souborů:",
+  "app.editEnvironmentConfig.variablesInfo": "Tyto proměnné určují předpis souborů, které je možné odevzdat jako řešení, a jak jsou tyto soubory použity ve vstupech pipelines. Každá hodnota může obsahovat buď název souboru nebo wildcard (např. <code>solution.cpp</code>, <code>*.py</code>, <code>my-*.\\{c,h\\}</code>). Povolené jsou pouze proměnné typu <code>file[]</code>.",
   "app.editEnvironmentSimpleForm.submit": "Uložit prostředí",
   "app.editEnvironmentSimpleForm.submitting": "Ukládám ...",
   "app.editEnvironmentSimpleForm.success": "Konfigurace byla změněna.",
@@ -191,13 +204,15 @@
   "app.editExercise.deleteExerciseWarning": "Smazání úlohy odstraní všechna studentská řešení a všechna zadání této úlohy.",
   "app.editExercise.description": "Změna nastavení úlohy",
   "app.editExercise.editConfig": "Nastavení konfigurace úlohy",
-  "app.editExercise.runtimeEnvironment": "Runtime Environment",
-  "app.editExercise.runtimeEnvironments": "Běhová prostředí",
-  "app.editExercise.testsAndScoring": "Testy a bodování",
   "app.editExercise.title": "Změna nastavení úlohy",
   "app.editExerciseConfig.changeConfigAdvancedExplain": "Po změně na pokročilou konfiguraci bude možné použít libovolné pipelines a ručně nakonfigurovat jejich parametry. Úloha bude také moci definovat vlastní omezení, jaké soubory smí studenti odevzdávat. Pro správné nastavení pokročilé konfigurace je ale zapotřebí hlubší vhled do vyhodnocovacího procesu ReCodExu. Pokročilá konfigurace může mít nastavené pouze jedno běhové prostředí, takže není možné změnit typ konfigurace, pokud má úloha nastaveno více běhových prostředí.",
   "app.editExerciseConfig.changeConfigSimpleExplain": "Návrat k jednoduché konfiguraci vyžaduje, aby úloha neměla žádné běhové prostředí (a tedy ani konfiguraci), protože není možné konvertovat libovolnou kombinaci pipeline tak, aby výsledek vyhovoval omezením formuláře pro jednoduchou konfiguraci.",
   "app.editExerciseConfig.description": "Změnit nastavení testů úlohy",
+  "app.editExerciseConfig.editConfig": "Editovat nastavení úlohy",
+  "app.editExerciseConfig.noRuntimes": "Běhová prostředí úlohy ještě nebyla nastavena. Prostředí musí být nastavena, aby bylo možné pracovat s konfigurací pipelines.",
+  "app.editExerciseConfig.noTests": "Nebyly nastaveny žádné testy. Formulář není možné zobrazit, dokud nebude alespoň jeden test nastaven.",
+  "app.editExerciseConfig.runtimeEnvironments": "Běhová prostředí",
+  "app.editExerciseConfig.testsAndScoring": "Testy a bodování",
   "app.editExerciseConfig.title": "Změnit nastavení testů",
   "app.editExerciseConfigForm.smartFillAll": "Chytře vyplnit vše",
   "app.editExerciseConfigForm.smartFillAll.yesNoQuestion": "Opravdu chcete přepsat konfiguraci všech následných testů použitím prvního testu jako šablony? Soubory budou spárovány k jednotlivým konfiguracím testů heuristikou založenou na odpovídajících podřetězcích.",
