+++ src/locales/en.json
@@ -182,6 +182,19 @@
   "app.editAssignmentForm.validation.pointsPercentualThresholdMustBeInteger": "Points percentual threshold must be an integer.",
   "app.editAssignmentForm.validation.sameLocalizedTexts": "There are more language variants with the same locale. Please make sure locales are unique.",
   "app.editAssignmentForm.validation.submissionsCountLimit": "Please fill the submissions count limit field with a positive integer.",
+  "app.editEnvironmentConfig.noRuntimeSelected": "There must be a runtime environment selected before you can proceed with exercise configuration.",
+  "app.editEnvironmentConfig.noVariables": "There are no variables yet ...",
+  "app.editEnvironmentConfig.selectedRuntime": "Selected Runtime Environment:",
+  "app.editEnvironmentConfig.selectedRuntimeInfo": "In the advanced configuration, selected runtime environemnt is used only to ensure that the backend worker has necessary compilers, tools, or libraries required by the environment. Everything else (source files patterns, pipelines) is configured separately.",
+  "app.editEnvironmentConfig.setDefaultVariables": "Set Default Variables",
+  "app.editEnvironmentConfig.title": "Runtime Environment Configuration",
+  "app.editEnvironmentConfig.tooltip.add": "Add another variable.",
+  "app.editEnvironmentConfig.tooltip.remove": "Remove this variable.",
+  "app.editEnvironmentConfig.validateName": "This is not a valid variable name.",
+  "app.editEnvironmentConfig.validateWildcard": "This value is not a valid file name or wildcard.",
+  "app.editEnvironmentConfig.variableName": "Source Files Variable:",
+  "app.editEnvironmentConfig.variableValue": "Wildcard Pattern:",
+  "app.editEnvironmentConfig.variablesInfo": "These variables cover the submitted files and how they are associated with pipeline inputs. Each value may hold a file name or a wildcard (e.g., <code>solution.cpp</code>, <code>*.py</code>, <code>my-*.\\{c,h\\}</code>). Only variables of type <code>file[]</code> are allowed here.",
   "app.editEnvironmentSimpleForm.submit": "Save Environments",
   "app.editEnvironmentSimpleForm.submitting": "Saving Environments ...",
   "app.editEnvironmentSimpleForm.success": "Environments Saved.",
@@ -191,13 +204,15 @@
   "app.editExercise.deleteExerciseWarning": "Deleting an exercise will remove all the students submissions and all assignments.",
   "app.editExercise.description": "Change exercise settings",
   "app.editExercise.editConfig": "Edit exercise configuration",
-  "app.editExercise.runtimeEnvironment": "Runtime Environment",
-  "app.editExercise.runtimeEnvironments": "Runtime Environments",
-  "app.editExercise.testsAndScoring": "Tests and Scoring",
   "app.editExercise.title": "Edit exercise settings",
   "app.editExerciseConfig.changeConfigAdvancedExplain": "Changing to advanced configuration will allow you to use custom pipelines and manually configure their parameters. The exercise may also have a custom environment configuration alowing arbitrary constraints on which files are submitted by the students. A deeper understanding of the ReCodEx evaluation process is required to set the configuration correctly. Please note that advanced configurations may have only one runtime environment (so it is not possible to convert simple configurations with multiple environments set).",
   "app.editExerciseConfig.changeConfigSimpleExplain": "Changing back to simple configuration requires that the exercise have no runtime environment (and thus no configuration) set since it is not possible to convert advanced configuration with custom pipelines into simple configuration.",
   "app.editExerciseConfig.description": "Change exercise tests configuration",
+  "app.editExerciseConfig.editConfig": "Edit exercise configuration",
+  "app.editExerciseConfig.noRuntimes": "The runtime environment is not properly configured yet. A runtime must be selected before pipeline configuration becomes available.",
+  "app.editExerciseConfig.noTests": "There are no tests yet. The form cannot be displayed until at least one test is created.",
+  "app.editExerciseConfig.runtimeEnvironments": "Runtime Environments",
+  "app.editExerciseConfig.testsAndScoring": "Tests and Scoring",
   "app.editExerciseConfig.title": "Edit tests configuration",
   "app.editExerciseConfigForm.smartFillAll": "Smart Fill All",
   "app.editExerciseConfigForm.smartFillAll.yesNoQuestion": "Do you really wish to overwrite configuration of all subsequent tests using the first test as a template? Files will be paired to individual test configurations by a heuristics based on matching name substrings.",
